revert platform dependent binding
