+++ p256k1/src/context.rs
@@ -1,4 +1,4 @@
-use crate::_bindings::{secp256k1_context, SECP256K1_CONTEXT_SIGN};
+use crate::bindings::{secp256k1_context, SECP256K1_CONTEXT_SIGN};
 
 use super::_rename::{secp256k1_context_create, secp256k1_context_destroy};
