+++ p256k1/src/ecdsa.rs
@@ -1,4 +1,4 @@
-use crate::_bindings::{secp256k1_ecdsa_signature, secp256k1_pubkey, SECP256K1_EC_COMPRESSED};
+use crate::bindings::{secp256k1_ecdsa_signature, secp256k1_pubkey, SECP256K1_EC_COMPRESSED};
 use crate::_rename::{
     secp256k1_ec_pubkey_create, secp256k1_ec_pubkey_parse, secp256k1_ec_pubkey_serialize,
     secp256k1_ecdsa_sign, secp256k1_ecdsa_signature_parse_compact,
