+++ p256k1/src/scalar.rs
@@ -10,7 +10,7 @@ use core::{
 use num_traits::{One, Zero};
 use rand_core::{CryptoRng, RngCore};
 
-use crate::_bindings::secp256k1_scalar;
+use crate::bindings::secp256k1_scalar;
 use crate::_rename::{
     secp256k1_ecmult, secp256k1_scalar_add, secp256k1_scalar_eq, secp256k1_scalar_get_b32,
     secp256k1_scalar_inverse, secp256k1_scalar_mul, secp256k1_scalar_negate,
