+++ update/Cargo.toml
@@ -11,4 +11,3 @@ syn = { version = "2.0.10", features = ["full"] }
 itertools = "0.10.5"
 proc-macro2 = "1.0.54"
 rustfmt-wrapper = "0.2.0"
-quote = "1.0.21"
